CountingSort
//https://www.faceprep.in/python/counting-sort/

//easy

//counting sort

/* Time Complexity
O(2n + (l - k)) where n : size of input, k : min of the input, l : max of the input.
2n : iterated once to get max and min, iterated once to add in the count array
l - k : this is the size of count array iterated through it once.
 */

/* Space Complexity
O(l - k) where k : min of the input, l : max of the input.
l - k : this is the size of count array
 */

//Goldman Sachs Practice